server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
google-genai==0.8.0
ollama==0.3.3
langdetect==1.0.9
bcrypt==4.2.0
slowapi==0.1.9
pydantic==2.8.2
httpx==0.27.2